Optimise texture updates on camera angle change for improved performance. Use new Luabind CMake file. Use new Xerces-C CMake file. Solve signed/unsigned comparison warning in SpatialContainer1D. Solve variable name reuse warning in DOMNodeWrapper. Use wide-screen windowed mode by default.
git-svn-id: svn://svn.code.sf.net/p/spindizzyremake/code@323 eb6eb719-311c-42bc-8795-40e0810bf6b9

+++ IsoRealms/Struct/SpatialContainer1D.h
@@ -70,7 +70,7 @@ template <class T> class SpatialContainer1D {
       cCells.insert(cCells.begin(), mCell);
      
     // If the object is "after" the last cell
-    } else if (mCellIndex >= cCells.size()) {
+    } else if (mCellIndex >= static_cast<int>(cCells.size())) {
       for (int i = cCells.size(); i < mCellIndex; i++) {
         cCells.push_back(nullptr);
       }
